:truck: :recycle: refactor: encapsulate bottom sheet in a package
